(wip): saving progress

+++ src/services/collectionsService.ts
@@ -1,23 +1,18 @@
 import { CollectionInsert, CollectionUpdate } from '@/utils/types'
 
-import { AuthenticationService } from './authenticationService'
 import { ServiceLocator } from './serviceLocator'
 import { ICollectionsRepository } from '@/repositories'
 import { CollectionDTO } from '@/shared/dtos/collection'
+import { UnauthorizedError } from '@/shared/errors/authErrors'
 
 export class CollectionsService {
-  private _collectionsRepository: ICollectionsRepository
-  private _authenticationService: AuthenticationService
-
-  constructor(collectionsRepository: ICollectionsRepository) {
-    this._collectionsRepository = collectionsRepository
-    this._authenticationService = ServiceLocator.getService(
-      'AuthenticationService'
-    )
-  }
+  constructor(private _collectionsRepository: ICollectionsRepository) {}
 
   async createCollection(collection: CollectionInsert): Promise<CollectionDTO> {
-    const user = await this._authenticationService.getUser()
+    const authenticationService = ServiceLocator.getService(
+      'AuthenticationService'
+    )
+    const user = await authenticationService.getUser()
 
     const newCollection = await this._collectionsRepository.createCollection(
       collection,
@@ -27,14 +22,25 @@ export class CollectionsService {
     return newCollection
   }
 
-  async getCollectionForUser(fingerprint: string): Promise<CollectionDTO> {
-    const user = await this._authenticationService.getUser()
+  async getPublicCollection(fingerprint: string): Promise<CollectionDTO> {
+    const collection =
+      await this._collectionsRepository.getCollection(fingerprint)
+    return collection
+  }
+
+  async getCollection(fingerprint: string): Promise<CollectionDTO> {
+    const authenticationService = ServiceLocator.getService(
+      'AuthenticationService'
+    )
+    const user = await authenticationService.getUser()
 
     const collection =
       await this._collectionsRepository.getCollection(fingerprint)
 
     if (collection.created_by !== user.id) {
-      throw new Error('Bad Request: collection does not belong to user')
+      throw new UnauthorizedError(
+        'Bad Request: collection does not belong to user'
+      )
     }
 
     return collection
@@ -44,7 +50,17 @@ export class CollectionsService {
     fingerprint: string,
     input: Partial<CollectionUpdate>
   ): Promise<CollectionDTO> {
-    const collection = await this.getCollectionForUser(fingerprint)
+    const authenticationService = ServiceLocator.getService(
+      'AuthenticationService'
+    )
+    const user = await authenticationService.getUser()
+    const collection = await this.getCollection(fingerprint)
+
+    if (collection.created_by !== user.id) {
+      throw new UnauthorizedError(
+        'Cannot update collection. Reason: unauthorized.'
+      )
+    }
 
     const newData: CollectionUpdate = {
       title: input.title ?? collection.title,
@@ -59,7 +75,17 @@ export class CollectionsService {
   }
 
   async deleteCollection(fingerprint: string): Promise<CollectionDTO> {
-    const collection = await this.getCollectionForUser(fingerprint)
+    const authenticationService = ServiceLocator.getService(
+      'AuthenticationService'
+    )
+    const user = await authenticationService.getUser()
+    const collection = await this.getCollection(fingerprint)
+
+    if (collection.created_by !== user.id) {
+      throw new UnauthorizedError(
+        'Cannot delete collection. Reason: unauthorized.'
+      )
+    }
 
     const deletedCollection =
       await this._collectionsRepository.deleteCollection(collection.fingerprint)
